Create data_exploration.ipynb
Created an initial data exploration file which loads the data.

This will be used to load data from the labelled planets csv and filters the confirmed planets and false positives. Then we will explore a the data to potentially identify some features that we can focus on with our ML model